+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.0.6
+
+- Add the `presets.excludeProps` setting to tell the options that some properties are not handled by the presets
+- Enhance display
+- Add possibility to specify `.html` files for presets preview. Will be loaded inside an `iframe`
+
 ## 0.0.5
 
 - Add the `debug.togglePropsKey` setting to allow force props display using `cmd + shift + p`